refs #48 Add page.js and initial client side routes
