+++ client/fixtures/typeahead.json
@@ -1,0 +1,282 @@
+[
+  "Ada Lovelace",
+  "Agnes Arber",
+  "Ahmed Zewail",
+  "Al-Battani",
+  "Alan Turing",
+  "Albert Abraham Michelson",
+  "Albert Einstein",
+  "Alberto Santos-Dumont",
+  "Albrecht von Haller",
+  "Aldo Leopold",
+  "Alessandro Volta",
+  "Alexander Bain",
+  "Alexander Fleming",
+  "Alexander Graham Bell",
+  "Alexander Von Humboldt",
+  "Alfred Binet",
+  "Alfred Kinsey",
+  "Alfred Nobel",
+  "Alfred Russel Wallace",
+  "Alfred Wegener",
+  "Amedeo Avogadro",
+  "Anaximander",
+  "Anders Celsius",
+  "André Marie Ampère",
+  "Andreas Vesalius",
+  "Angel Alcala",
+  "Antoine Lavoisier",
+  "Antonie van Leeuwenhoek",
+  "Antonio Meucci",
+  "Archimedes",
+  "Aristarchus",
+  "Aristotle",
+  "Arnold Sommerfeld",
+  "Arthur Eddington",
+  "Avicenna",
+  "B. F. Skinner",
+  "Barbara McClintock",
+  "Beatrix Potter",
+  "Benjamin Franklin",
+  "Benjamin Thompson",
+  "Bernardo Houssay",
+  "Bill Nye",
+  "Blaise Pascal",
+  "Brahmagupta",
+  "Brian Cox",
+  "C. V. Raman",
+  "Carl Bosch",
+  "Carl Friedrich Gauss",
+  "Carl Sagan",
+  "Carolus Linnaeus",
+  "Charles Babbage",
+  "Charles Darwin",
+  "Charles Lyell",
+  "Charles Sherrington",
+  "Charles-Augustin de Coulomb",
+  "Chen-Ning Yang",
+  "Christiaan Huygens",
+  "Christiane Nusslein-Volhard",
+  "Clarence Birdseye",
+  "Claude Bernard",
+  "Claude Levi-Strauss",
+  "Clyde Tombaugh",
+  "Daniel Bernoulli",
+  "David Bohm",
+  "David Hilbert",
+  "Dian Fossey",
+  "Dmitri Mendeleev",
+  "Dorothy Hodgkin",
+  "E. O. Wilson",
+  "Edmund Halley",
+  "Edward Jenner",
+  "Edward Teller",
+  "Edwin Herbert Land",
+  "Edwin Hubble",
+  "Elizabeth Blackwell",
+  "Emil Fischer",
+  "Emil Kraepelin",
+  "Emile Berliner",
+  "Emmy Noether",
+  "Empedocles",
+  "Enrico Fermi",
+  "Eratosthenes",
+  "Ernest Lawrence",
+  "Ernest Rutherford",
+  "Ernst Haeckel",
+  "Ernst Mach",
+  "Ernst Mayr",
+  "Ernst Werner von Siemens",
+  "Erwin Chargaff",
+  "Erwin Schrodinger",
+  "Euclid",
+  "Evangelista Torricelli",
+  "Fibonacci",
+  "Francesco Redi",
+  "Francis Bacon",
+  "Francis Crick",
+  "Francis Galton",
+  "Franz Boas",
+  "Frederick Gowland Hopkins",
+  "Frederick Sanger",
+  "Frederick Soddy",
+  "Friedrich August Kekulé",
+  "Friedrich Wöhler",
+  "Fritz Haber",
+  "Galen",
+  "Galileo Galilei",
+  "Gene Shoemaker",
+  "Georg Ohm",
+  "George Gaylord Simpson",
+  "George Washington Carver",
+  "Georges-Louis Leclerc, Comte de Buffon",
+  "Gertrude Elion",
+  "Gerty Theresa Cori",
+  "Glenn Seaborg",
+  "Gottfried Leibniz",
+  "Gottlieb Daimler",
+  "Grace Murray Hopper",
+  "Gregor Mendel",
+  "Guglielmo Marconi",
+  "Gustav Kirchoff",
+  "Gustav Ludwig Hertz",
+  "Hans Christian Oersted",
+  "Hans Selye",
+  "Harold Urey",
+  "Harriet Quimby",
+  "Hedy Lamarr",
+  "Heinrich Hertz",
+  "Henri Becquerel",
+  "Henrietta Swan Leavitt",
+  "Henry Bessemer",
+  "Henry Cavendish",
+  "Henry Ford",
+  "Henry Moseley",
+  "Hermann Rorschach",
+  "Hermann von Helmholtz",
+  "Homi Jehangir Bhabha",
+  "Humphry Davy",
+  "Ibn Rushd",
+  "Inge Lehmann",
+  "Irene Joliot-Curie",
+  "Isaac Newton",
+  "Ivan Pavlov",
+  "J. J. Thomson",
+  "J. Robert Oppenheimer",
+  "J. Willard Gibbs",
+  "Jack Horner",
+  "Jacques Cousteau",
+  "Jagadish Chandra Bose",
+  "James Chadwick",
+  "James Clerk Maxwell",
+  "James Dwight Dana",
+  "James Hutton",
+  "James Prescott Joule",
+  "James Watson",
+  "James Watt",
+  "Jan Baptist von Helmont",
+  "Jane Goodall",
+  "Jane Marcet",
+  "Jean Piaget",
+  "Jean-Baptiste Lamarck",
+  "Jocelyn Bell Burnell",
+  "Johannes Kepler",
+  "John Dalton",
+  "John Locke",
+  "John Logie Baird",
+  "John Napier",
+  "John Needham",
+  "John Ray",
+  "John von Neumann",
+  "Jonas Salk",
+  "Joseph Banks",
+  "Joseph Henry",
+  "Joseph Lister",
+  "Joseph Priestley",
+  "Justus von Liebig",
+  "Karl Landsteiner",
+  "Katharine Burr Blodgett",
+  "Kip S. Thorne",
+  "Konrad Lorenz",
+  "Lee De Forest",
+  "Leonardo da Vinci",
+  "Leonhard Euler",
+  "Lester R. Brown",
+  "Linus Pauling",
+  "Lise Meitner",
+  "Louis Agassiz",
+  "Louis de Broglie",
+  "Louis Pasteur",
+  "Lucretius",
+  "Ludwig Boltzmann",
+  "Luigi Galvani",
+  "Luis Alvarez",
+  "Luther Burbank",
+  "Lynn Margulis",
+  "Mae Carol Jemison",
+  "Marcello Malpighi",
+  "Marguerite Perey",
+  "Maria Gaetana Agnesi",
+  "Maria Goeppert-Mayer",
+  "Maria Mitchell",
+  "Marie Curie",
+  "Mario Molina",
+  "Mary Anning",
+  "Max Born",
+  "Max Planck",
+  "Max von Laue",
+  "Michael E. Brown",
+  "Michael Faraday",
+  "Michio Kaku",
+  "Mohammad Abdus Salam",
+  "Muhammad ibn Musa al-Khwarizmi",
+  "Murray Gell-Mann",
+  "Nicolaus Copernicus",
+  "Niels Bohr",
+  "Nikola Tesla",
+  "Noam Chomsky",
+  "Omar Khayyam",
+  "Otto Hahn",
+  "Paul Dirac",
+  "Paul Ehrlich",
+  "Pierre Curie",
+  "Pierre de Fermat",
+  "Pierre-Simon Laplace",
+  "Prafulla Chandra Ray",
+  "Pythagoras",
+  "Rachel Carson",
+  "René Descartes",
+  "Richard Feynman",
+  "Rita Levi-Montalcini",
+  "Robert Bosch",
+  "Robert Boyle",
+  "Robert Brown",
+  "Robert Bunsen",
+  "Robert Goddard",
+  "Robert Hooke",
+  "Robert Koch",
+  "Ronald Fisher",
+  "Ronald Ross",
+  "Rosalind Franklin",
+  "Rudolf Christian Karl Diesel",
+  "Rudolf Virchow",
+  "Salim Ali",
+  "Sally Ride",
+  "Santiago Ramón y Cajal",
+  "Sergei Winogradsky",
+  "Sigmund Freud",
+  "Srinivasa Ramanujan",
+  "Stephanie Kwolek",
+  "Stephen Hawking",
+  "Subrahmanyan Chandrasekhar",
+  "Svante Arrhenius",
+  "Thabit ibn Qurra",
+  "Thales",
+  "Theodor Schwann",
+  "Theodosius Dobzhansky",
+  "Thomas Alva Edison",
+  "Thomas Hunt Morgan",
+  "Thomas Kuhn",
+  "Thomas Newcomen",
+  "Thomas Willis",
+  "Tim Noakes",
+  "Timothy John Berners-Lee",
+  "Tycho Brahe",
+  "Virginia Apgar",
+  "Vladimir Vernadsky",
+  "Werner Heisenberg",
+  "Wernher Von Braun",
+  "Wilbur and Orville Wright",
+  "Wilhelm Conrad Roentgen",
+  "Wilhelm Röntgen",
+  "Wilhelm Wundt",
+  "William Buckland",
+  "William Harvey",
+  "William Herschel",
+  "William Hopkins",
+  "William Ramsay",
+  "William Smith",
+  "William Thomson",
+  "Wolfgang Ernst Pauli",
+  "Zora Neale Hurston"
+]
