+++ client/middleware/error-404.js
@@ -1,0 +1,12 @@
+module.exports = function (page) {
+  page(function (ctx, next) {
+    console.warn('Page not found', ctx.path);
+
+    // Temporary until templates pulled into js for client side rendering
+    if (!ctx.isInitialRender) {
+      window.location = ctx.path;
+    }
+
+    next();
+  });
+};
