+++ client/middleware/initial-render.js
@@ -1,0 +1,9 @@
+var isInitialRender = true;
+
+module.exports = function (page) {
+  page(function (ctx, next) {
+    ctx.isInitialRender = isInitialRender;
+    isInitialRender = false;
+    next();
+  });
+};
